requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Limiter==3.8.0
psycopg2-binary==2.9.9
Werkzeug==3.0.3
alembic==1.13.1
gunicorn==22.0.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
